LayoutTests/fast/frames/iframe-no-name.html

<!DOCTYPE html>
<html>
<head>
<link rel="help" href="http://www.whatwg.org/specs/web-apps/current-work/multipage/the-iframe-element.html#attr-iframe-name">
<script src="../../resources/js-test.js"></script>
</head>
<body>
<iframe id="id"></iframe>
<iframe name="name"></iframe>
<iframe id="id" name="name"></iframe>
<iframe name="name" id="id"></iframe>
<script>
description("Checks that the id of an iframe does not set the contentWindow's name if the iframe's name is not set.");
shouldBeEqualToString("frames[0].name", "");
shouldBeEqualToString("frames[1].name", "name");
shouldBeEqualToString("frames[2].name", "name");
shouldBeEqualToString("frames[3].name", "name");
</script>
</body>
</html>
