LayoutTests/fast/frames/iframe-no-name-expected.txt

Checks that the id of an iframe does not set the contentWindow's name if the iframe's name is not set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS frames[0].name is ""
PASS frames[1].name is "name"
PASS frames[2].name is "name"
PASS frames[3].name is "name"
PASS successfullyParsed is true

TEST COMPLETE
